LoCMapBot/requirements.txt
python-twitter == 3.3
pandas == 0.19.1
xmltodict == 0.11.0